suppr user

+++ Src/Controleur/del_userControleur.php
@@ -8,17 +8,22 @@
 
         public function suppressionUtilisateur(Request $request){
             $data = $request->getData();
-            var_dump($data);
-            $u = UtilisateurDAO::getByLogin($data['name_user']);
-            if($u !== false)
-            {
-                UtilisateurDAO::delete($u);
-                header("Location: user");
-                exit;
+            if ($data["choix"] == "oui"){
+                $u = UtilisateurDAO::getById($request->getId());
+                if($u !== false)
+                {
+                    UtilisateurDAO::delete($u);
+                    header("Location: /user");
+                    exit;
+                }
+                else{
+                    $_SESSION["delUser"] = false;
+                    header("Location: /delUser/" . $request->getId());
+                    exit;
+                }
             }
             else{
-                $_SESSION["delUser"] = false;
-                header("Location: /delUser");
+                header("Location: /user");
                 exit;
             }
         }
